requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch
torchvision
torchaudio
notebook
numpy
pandas
matplotlib
seaborn
python-box==6.0.2
pyYaml
tqdm
joblib
types-PyYAML
gdown
-e .